added distinct message for question add & update. Merged all utils in a single file
