--- pkg/utils/uuid.go
@@ -1,7 +1,0 @@
-package utils
-
-import "github.com/google/uuid"
-
-func GenerateUUID() string {
-	return uuid.New().String()
-}
